refactor: extract builder shared primitives
