+++ tests/integration/synthetic.ts
@@ -17,9 +17,11 @@ import {
 	happinessTier,
 	effect,
 	passiveParams,
+} from '@kingdom-builder/contents/config/builders';
+import {
 	Types,
 	PassiveMethods,
-} from '@kingdom-builder/contents/config/builders';
+} from '@kingdom-builder/contents/config/builderShared';
 
 export function createSyntheticContext() {
 	const costKey = 'r0';
